fix: node.js module import error when using middleware
